+++ tests/plugin_test/src/199_never_type_and_plugins.php
@@ -3,7 +3,7 @@
 function exitMisuse(?int $x, bool $z): array {
     $x1 = $x ?: exit('expected nonzero');
     $z2 = $z || exit(true);
-    $y = $z ?: exit('expected nonzero');
+    $y = $z ?: exit('expected nonzero');  // above check implies $z is true
     $y3 = $x ?? exit('fail');
     '@phan-debug-var $y, $x1, $z2, $y3'; // Not worth specializing $z2 right now, using the return type would be uncommon
     $y2 = $x + exit('fail');